Add support for Xcode 12.5 beta

+++ Plugins/Shared Code/Xcode/IDEOpenRequest-Protocol.h
@@ -10,7 +10,7 @@
 
 
 @protocol IDEOpenRequest <NSObject>
-@property(readonly) int requestState;
+@property(readonly) unsigned long long requestState;
 - (void)cancel;
 @end
